+++ backend/utils/variable_filler.go
@@ -88,3 +88,99 @@ func TemplateVars_Autodownload(mediaItem models.MediaItem, setItem models.DBPost
 		},
 	)
 }
+
+func TemplateVars_DownloadQueue(mediaItem models.MediaItem, setItem models.DBPosterSetDetail, Errors []string, Warnings []string) map[string]string {
+	var result string
+	if len(Errors) > 0 {
+		result = "Error"
+	} else if len(Warnings) > 0 {
+		result = "Warning"
+	} else {
+		result = "Success"
+	}
+	reasonTitle := result
+	var reason string
+	switch result {
+	case "Success":
+		reason = "Download completed successfully with no issues detected."
+	case "Warning":
+		reason = fmt.Sprintf("Download completed with warnings:\n%s", strings.Join(Warnings, "\n"))
+	case "Error":
+		reason = fmt.Sprintf("Download completed with errors:\n%s", strings.Join(Errors, "\n"))
+		if len(Warnings) > 0 {
+			reason += fmt.Sprintf("\n\n and the following Warnings:\n%s", strings.Join(Warnings, "\n"))
+		}
+	}
+	return MergeTemplateVars(
+		BaseTemplateVars(),
+		map[string]string{
+			"MediaItemTitle":        mediaItem.Title,
+			"MediaItemYear":         fmt.Sprintf("%d", mediaItem.Year),
+			"MediaItemTMDBID":       mediaItem.TMDB_ID,
+			"MediaItemLibraryTitle": mediaItem.LibraryTitle,
+			"MediaItemRatingKey":    mediaItem.RatingKey,
+			"MediaItemType":         mediaItem.Type,
+			"SetID":                 setItem.ID,
+			"SetTitle":              setItem.Title,
+			"SetType":               setItem.Type,
+			"SetCreator":            setItem.UserCreated,
+			"ReasonTitle":           reasonTitle,
+			"Reason":                reason,
+		},
+	)
+}
+
+func TemplateVars_NewSetsAvailableForIgnoredItems(mediaItem models.MediaItem, setCount int) map[string]string {
+	return MergeTemplateVars(
+		BaseTemplateVars(),
+		map[string]string{
+			"MediaItemTitle":        mediaItem.Title,
+			"MediaItemYear":         fmt.Sprintf("%d", mediaItem.Year),
+			"MediaItemTMDBID":       mediaItem.TMDB_ID,
+			"MediaItemLibraryTitle": mediaItem.LibraryTitle,
+			"MediaItemRatingKey":    mediaItem.RatingKey,
+			"MediaItemType":         mediaItem.Type,
+			"SetCount":              fmt.Sprintf("%d", setCount),
+		},
+	)
+}
+
+func TemplateVars_CheckForMediaItemChangesJob(mediaItem models.MediaItem, reason string, action string, moreInfo string) map[string]string {
+	return MergeTemplateVars(
+		BaseTemplateVars(),
+		map[string]string{
+			"MediaItemTitle":        mediaItem.Title,
+			"MediaItemYear":         fmt.Sprintf("%d", mediaItem.Year),
+			"MediaItemTMDBID":       mediaItem.TMDB_ID,
+			"MediaItemLibraryTitle": mediaItem.LibraryTitle,
+			"MediaItemRatingKey":    mediaItem.RatingKey,
+			"MediaItemType":         mediaItem.Type,
+			"Reason":                reason,
+			"Action":                action,
+			"MoreInfo":              moreInfo,
+		},
+	)
+}
+
+func TemplateVars_SonarrNotification(mediaItem models.MediaItem, setItem models.DBPosterSetDetail, image models.ImageFile, reasonTitle string, reason string, result string) map[string]string {
+	return MergeTemplateVars(
+		BaseTemplateVars(),
+		map[string]string{
+			"MediaItemTitle":        mediaItem.Title,
+			"MediaItemYear":         fmt.Sprintf("%d", mediaItem.Year),
+			"MediaItemTMDBID":       mediaItem.TMDB_ID,
+			"MediaItemLibraryTitle": mediaItem.LibraryTitle,
+			"MediaItemRatingKey":    mediaItem.RatingKey,
+			"MediaItemType":         mediaItem.Type,
+			"SetID":                 setItem.ID,
+			"SetTitle":              setItem.Title,
+			"SetType":               setItem.Type,
+			"SetCreator":            setItem.UserCreated,
+			"ImageName":             GetFileDownloadName(mediaItem.Title, image),
+			"ImageType":             image.Type,
+			"ReasonTitle":           reasonTitle,
+			"Reason":                reason,
+			"Result":                result,
+		},
+	)
+}